neozzzzzzzzz

+++ ft_strlcat.c
@@ -6,11 +6,13 @@
 /*   By: pflorent <pflorent@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/08 16:43:40 by pflorent          #+#    #+#             */
-/*   Updated: 2021/01/08 17:24:20 by pflorent         ###   ########.fr       */
+/*   Updated: 2021/01/09 16:28:07 by pflorent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlcat(char *dest, const char *src, int lgt)
+# include "libft.h"
+
+size_t		ft_strlcat(char *dest, const char *src, int lgt)
 {
 	int dest_size;
 	int i;
