+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: pflorent <pflorent@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/09 15:42:59 by pflorent          #+#    #+#             */
-/*   Updated: 2021/01/09 15:45:06 by pflorent         ###   ########.fr       */
+/*   Updated: 2021/01/09 16:28:27 by pflorent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	*ft_memccpy(void *dest, const void *src, int ch, size_t size);
 void	*ft_memmove (void *dest, const void *src, size_t size);
 void    *ft_memchr (const void *mem_block, int ch, size_t size);
 int     ft_memcmp (const void *p1, const void *p2, size_t size);
-int     ft_strlen(char *str);
+size_t     ft_strlen(char *str);
 int     ft_isalpha(int a);
 int     ft_isdigit(int a);
 int     ft_isalnum(int a);
@@ -34,8 +34,8 @@ int     ft_tolower(int a);
 char    *ft_strchr(const char *str, int chr);
 char    *ft_strrchr(const char *str, int chr);
 int     ft_strncmp(const char *first_str, const char *second_str, size_t lgt);
-int     ft_strlcpy(char *dest, const char *src, int lgt);
-int     ft_strlcat(char *dest, const char *src, int lgt);
+size_t     ft_strlcpy(char *dest, const char *src, int lgt);
+size_t     ft_strlcat(char *dest, const char *src, int lgt);
 char    *ft_strnstr(const char *str, const char *find, unsigned int lgt);
 int     ft_atoi(const char *str);
 void    *ft_calloc(size_t elem_count, size_t elem_size);
